# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/influencers": {
            "get": {
                "tags": [
                    "Influencers"
                ],
                "summary": "List all influencers",
                "operationId": "770f62a4b258ab36753bb4d915ecf882",
                "responses": {
                    "200": {
                        "description": "List of influencers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Influencer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Create a new influencer",
                "operationId": "be1d7b9b0b6353dedb1736bb8de9cae5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "instagram_handle"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "instagram_handle": {
                                        "type": "string",
                                        "example": "@johndoe"
                                    },
                                    "followers_count": {
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Influencer created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/influencers/{id}": {
            "get": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Get a specific influencer",
                "operationId": "5d299bdb16bd6dc2d6944403dedc7e03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Influencer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Update an influencer",
                "operationId": "558bc79477c2111cfa900c1456cc0c6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "instagram_handle"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Name"
                                    },
                                    "instagram_handle": {
                                        "type": "string",
                                        "example": "@updatedhandle"
                                    },
                                    "followers_count": {
                                        "type": "integer",
                                        "example": 12000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Influencer updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Influencer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Influencer not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Influencers"
                ],
                "summary": "Delete an influencer",
                "operationId": "96063a54ea1677a0bc5ee96dfce6dc15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Influencer deleted"
                    },
                    "404": {
                        "description": "Influencer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Influencer": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "followers": {
                        "type": "integer",
                        "example": 10000
                    },
                    "platform": {
                        "type": "string",
                        "example": "Instagram"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira seu token Bearer para autenticação",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}